Refactor: 유저 삭제 기능(회원탈퇴) 리팩토링

+++ src/main/java/com/example/schedulerapp/service/UserServiceJPA.java
@@ -60,8 +60,12 @@ public class UserServiceJPA implements UserService {
     }
 
     @Override
-    public void deleteByIdUser(Long id) {
+    public void deleteByIdUser(Long id, Long userId) {
         User findUser = userRepository.findByIdOrElseThrow(id);
+
+        if (!findUser.getId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You can only delete your own account.");
+        }
 
         userRepository.delete(findUser);
     }
